requirements.txt
python-dotenv==1.0.1
django-storages==1.14.2
asgiref==3.7.2
dj-database-url==1.3.0
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.4.0
boto3==1.34.50
environs==11.0.0
Pillow==9.5.0